requirements.txt
numpy 
pandas
matplotlib
scikit-learn
seaborn
xgboost
joblib
tensorflow==2.15.0
keras==2.15.0
Flask
Flask-PyMongo
Flask-Login
Flask-WTF
Flask-Bcrypt
Flask-CORS
pymongo
python-dotenv
requests
WTForms
email-validator
dnspython
gunicorn